pyproject.toml
[project]
name = "diffex"
version = "v0.3.0"
description = "Differential expression report generator using edgeR and Quarto"
dependencies = ["typer==0.15.2", "pyyaml"]

[project.scripts]
diffex = "diffex.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["diffex"]

[tool.setuptools.package-data]
"diffex" = ["DiffEx.qmd"]
